move all middlewares from schema files to middlewares path
