+++ backend/schemas/review.ts
@@ -1,9 +1,6 @@
 import { z } from 'zod'
-import { NextFunction, Request, Response } from 'express'
-import { validateSchemaData } from './index'
-import ExpressError from '../lib/classes'
 
-export const reviewSchema = z
+const reviewSchema = z
     .object({
         body: z.string({ required_error: 'Body is required!', invalid_type_error: 'Body must be a string' }).nonempty(),
         rating: z
@@ -13,8 +10,4 @@ export const reviewSchema = z
     })
     .required()
 
-export const validateReview = (req: Request, res: Response, next: NextFunction) => {
-    const { errors } = validateSchemaData(req.body, reviewSchema)
-    if (errors) throw new ExpressError('', 400, errors)
-    else next()
-}
+export default reviewSchema
